+++ static/zwolle/src/Ampersand/Core/Concept.php
@@ -107,6 +107,12 @@ class Concept {
 	 * @var DatabaseTable
 	 */
 	private $mysqlConceptTable;
+    
+    /**
+     * @var Atom[] $atomCache array with atoms that exist in the concept (within database transaction)
+     * used to prevent unnecessary queries to check if atom is already in database
+     */
+    private $atomCache = array();
 	
 	/**
 	 * Concept constructor
@@ -128,7 +134,7 @@ class Concept {
 		    
 		    if ($conj->isSigConj()) $this->affectedSigConjuncts[] = $conj;
 		    if ($conj->isInvConj()) $this->affectedInvConjuncts[] = $conj;
-		    if (!$conj->isSigConj() && !$conj->isInvConj()) $this->logger->warning("Affected conjunct '{$conj->id}' (specified for concept '[{$this->name}]') is not part of an invariant or signal rule");
+		    // if (!$conj->isSigConj() && !$conj->isInvConj()) $this->logger->warning("Affected conjunct '{$conj->id}' (specified for concept '[{$this->name}]') is not part of an invariant or signal rule");
 		}
 		
 		$this->specializations = (array)$conceptDef['specializations'];
@@ -233,9 +239,13 @@ class Concept {
 	 * @return mixed[]
 	 */
 	public function getAllAtomObjects(){
-	    $arr = array();
-	    foreach ($this->getAllAtomIds() as $tgtAtomId){
-	        $tgtAtom = new Atom($tgtAtomId, $this->name);
+        // Query all atoms in table
+        $firstCol = current($this->mysqlConceptTable->getCols()); // We can query an arbitrary concept col for checking the existence of an atom
+	    $query = "SELECT DISTINCT *, `{$firstCol->name}` as `atomId` FROM `{$this->mysqlConceptTable->name}` WHERE `{$firstCol->name}` IS NOT NULL";
+        
+        $arr = array();
+	    foreach ((array)$this->database->Exe($query) as $row){
+	        $tgtAtom = new Atom($row['atomId'], $this->name, null, $row);
 	        $arr[] = $tgtAtom->getAtom();
 	    }
 	    return $arr;
@@ -330,6 +340,22 @@ class Concept {
 	public function createNewAtom(){
 	    return new Atom($this->createNewAtomId(), $this->name);
 	}
+    
+    /**
+     * @param Atom $atom check if atom exists in concept atom cache
+     * @return boolean
+     */
+    public function inAtomCache($atom){
+        return in_array($atom, $this->atomCache);
+    }
+    
+    /**
+     * @param Atom $atom atom to add to concept atom cache
+     * @return void
+     */
+    public function addToAtomCache($atom){
+        $this->atomCache[] = $atom;
+    }
 	
     /**********************************************************************************************
      *