+++ ReleaseNotes.md
@@ -1,8 +1,11 @@
 # Release notes of Ampersand
 
 ## current development 
+  * Bug fix: Issue with SQL query [Issue #152](https://github.com/AmpersandTarski/Ampersand/issues/152)
+  * Bug fix: minor issue with SQL query [Issue #436](https://github.com/AmpersandTarski/Ampersand/issues/436)
   * Bug fix: Nontermination of functional document generator. [Issue #231](https://github.com/AmpersandTarski/Ampersand/issues/231)
-  * SQL query performance improvement: [Issue #426](https://github.com/AmpersandTarski/Ampersand/issues/426)
+  * SQL query performance improvements: [Issue #426](https://github.com/AmpersandTarski/Ampersand/issues/426) and 
+    [Issue #217](https://github.com/AmpersandTarski/Ampersand/issues/217)
   
 ## v3.5.1 (17 may 2016)
   * Minor enhancement of generation of Logical Data Model